# server/docs/swagger.yaml
definitions:
...
  nfc_relay_admin.BatchDecryptRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/nfc_relay_admin.DecryptAndVerifyRequest'
        type: array
      mode:
        description: PARALLEL, SEQUENTIAL
        type: string
    required:
    - items
    type: object
  nfc_relay_admin.BatchDecryptResponse:
    properties:
      failCount:
        type: integer
      results:
        items:
          $ref: '#/definitions/nfc_relay_admin.DecryptAndVerifyResponse'
        type: array
      successCount:
        type: integer
      totalTime:
        description: 微秒
        type: integer
    type: object
...
  nfc_relay_admin.DecryptAndVerifyRequest:
    properties:
      apduData:
        $ref: '#/definitions/security.APDUDataClass'
      auditLevel:
        description: BASIC, ENHANCED, STRICT
        type: string
      sessionId:
        type: string
      userId:
        type: string
      verifyOnly:
        description: 仅验证不解密
        type: boolean
    required:
    - apduData
    - sessionId
    - userId
    type: object
  nfc_relay_admin.DecryptAndVerifyResponse:
    properties:
      auditResult:
        $ref: '#/definitions/security.ComplianceResult'
      decryptedData: {}
      success:
        type: boolean
      verificationDetails:
        $ref: '#/definitions/nfc_relay_admin.VerificationDetails'
    type: object
...
  nfc_relay_admin.VerificationDetails:
    properties:
      complianceStatus:
        type: string
      decryptionTime:
        description: 微秒
        type: integer
      encryptionValid:
        type: boolean
      integrityValid:
        type: boolean
      riskAssessment:
        type: string
      verificationTime:
        description: 微秒
        type: integer
    type: object
...
  request.CreateAuditLogRequest:
    properties:
      action:
        description: 操作动作
        type: string
      category:
        description: 事件分类
        type: string
      client_id_initiator:
        description: 发起方客户端ID
        type: string
      client_id_responder:
        description: 响应方客户端ID
        type: string
      details:
        description: 事件详情
      duration:
        description: 操作耗时(毫秒)
        type: integer
      error_message:
        description: 错误信息
        type: string
      event_type:
        description: 事件类型
        type: string
      level:
        description: 日志级别
        type: string
      request_id:
        description: 请求ID
        type: string
      resource:
        description: 操作资源
        type: string
      result:
        description: 操作结果
        type: string
      server_id:
        description: 服务器ID
        type: string
      session_id:
        description: 会话ID
        type: string
      source_ip:
        description: 源IP地址
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        description: 用户ID
        type: string
    required:
    - event_type
    - level
    - result
    type: object
...
  response.AuditLogEntry:
    properties:
      action:
        description: 操作动作
        type: string
      category:
        description: 事件分类
        type: string
      client_id_initiator:
        description: 发起方客户端ID
        type: string
      client_id_responder:
        description: 响应方客户端ID
        type: string
      created_at:
        description: 创建时间
        type: string
      details:
        description: 事件详情JSON
        type: string
      duration:
        description: 操作耗时(毫秒)
        type: integer
      error_message:
        description: 错误信息
        type: string
      event_time:
        description: 事件发生时间
        type: string
      event_type:
        description: 事件类型
        type: string
      id:
        description: 日志ID
        type: integer
      level:
        description: 日志级别
        type: string
      request_id:
        description: 请求ID
        type: string
      resource:
        description: 操作资源
        type: string
      result:
        description: 操作结果
        type: string
      server_id:
        description: 服务器ID
        type: string
      session_id:
        description: 会话ID
        type: string
      source_ip:
        description: 源IP地址
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        description: 用户ID
        type: string
    type: object
...
  response.PaginatedAuditLogResponse:
    properties:
      list:
        description: 审计日志列表
        items:
          $ref: '#/definitions/response.AuditLogEntry'
        type: array
      page:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页条数
        type: integer
      total:
        description: 总数
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  security.APDUDataClass:
    properties:
      auditData:
        allOf:
        - $ref: '#/definitions/security.AuditableData'
        description: 审计数据 - 服务器可读，用于合规检查
      businessData:
        allOf:
        - $ref: '#/definitions/security.EncryptedBusinessData'
        description: 业务数据 - 服务器可解密，用于合规检查
      metadata:
        allOf:
        - $ref: '#/definitions/security.APDUMetadata'
        description: 元数据 - 明文，用于路由和监控
    type: object
  security.APDUMetadata:
    properties:
      checksumCrc:
        description: 数据完整性校验
        type: string
      clientId:
        type: string
      deviceInfo:
        description: 设备指纹
        type: string
      direction:
        description: '"command" 或 "response"'
        type: string
      sequenceNum:
        description: 序列号，防重放
        type: integer
      sessionId:
        type: string
      timestamp:
        type: string
      userId:
        type: string
    type: object
  security.AuditableData:
    properties:
      amount:
        description: 交易金额 (分为单位)
        type: integer
      applicationId:
        description: 应用ID (脱敏后)
        type: string
      commandClass:
        description: 命令类别 (SELECT, READ, WRITE等)
        type: string
      commandType:
        description: 具体命令类型
        type: string
      currency:
        description: 货币代码
        type: string
      merchantCategory:
        description: 商户类别
        type: string
      riskScore:
        description: 风险评分
        type: integer
      timestamp:
        description: 时间戳
        type: string
      transactionType:
        description: 交易类型
        type: string
    type: object
  security.ComplianceResult:
    properties:
      actions:
        description: 建议的处理动作
        items:
          type: string
        type: array
      compliant:
        type: boolean
      reason:
        type: string
      riskLevel:
        description: LOW, MEDIUM, HIGH, CRITICAL
        type: string
    type: object
  security.EncryptedBusinessData:
    properties:
      encryptedData:
        description: 所有业务数据都使用审计密钥加密，服务器可解密进行合规检查
        type: string
      encryptionInfo:
        allOf:
        - $ref: '#/definitions/security.EncryptionInfo'
        description: 加密元信息
    type: object
  security.EncryptionInfo:
    properties:
      algorithm:
        description: 加密算法
        type: string
      nonce:
        description: 随机数
        type: string
      tag:
        description: 认证标签
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.2
paths:
  /admin/nfc-relay/v1/audit-logs:
    get:
      consumes:
      - application/json
      description: 获取审计日志列表，支持分页和筛选
      parameters:
      - description: 事件分类
        in: query
        name: category
        type: string
      - description: 客户端ID
        in: query
        name: clientID
        type: string
      - description: 结束时间（ISO8601格式）
        in: query
        name: endTime
        type: string
      - description: 事件类型
        in: query
        name: eventType
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - description: 日志级别
        in: query
        name: level
        type: string
      - description: 页码
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: 操作结果
        in: query
        name: result
        type: string
      - description: 会话ID
        in: query
        name: sessionID
        type: string
      - description: 源IP
        in: query
        name: sourceIP
        type: string
      - description: 开始时间（ISO8601格式）
        in: query
        name: startTime
        type: string
      - description: 用户ID
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedAuditLogResponse'
              type: object
      summary: 获取审计日志列表
      tags:
      - NFC中继审计日志
    post:
      consumes:
      - application/json
      description: 创建新的审计日志记录
      parameters:
      - description: 创建审计日志
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateAuditLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建审计日志
      tags:
      - NFC中继审计日志
  /admin/nfc-relay/v1/audit-logs/batch:
    post:
      consumes:
      - application/json
      description: 批量创建多条审计日志记录
      parameters:
      - description: 批量创建审计日志
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/request.CreateAuditLogRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 批量创建审计日志
      tags:
      - NFC中继审计日志
  /admin/nfc-relay/v1/audit-logs/cleanup:
    delete:
      consumes:
      - application/json
      description: 删除指定天数之前的审计日志
      parameters:
      - description: 保留天数
        in: query
        name: retentionDays
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除过期审计日志
      tags:
      - NFC中继审计日志
...
  /admin/nfc-relay/v1/encryption/batch-decrypt-verify:
    post:
      consumes:
      - application/json
      description: 批量处理多个APDU数据的解密验证
      parameters:
      - description: 批量解密请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/nfc_relay_admin.BatchDecryptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.BatchDecryptResponse'
              type: object
      summary: 批量解密和验证APDU数据
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/encryption/decrypt-verify:
    post:
      consumes:
      - application/json
      description: 对接收到的APDU数据进行解密验证，支持合规检查
      parameters:
      - description: 解密验证请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/nfc_relay_admin.DecryptAndVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.DecryptAndVerifyResponse'
              type: object
      summary: 解密和验证APDU数据
      tags:
      - NFC中继管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...